init urlvr code

+++ verl/data/preprocess.py
@@ -34,13 +34,19 @@ def make_map_fn(split, source=None):
 
 if __name__ == '__main__':
 
-    data_source = 'MATH-L5-TTT'
+    data_sources = ['DAPO', 'AIME-TTT', 'AIME25-TTT', 'AMC-TTT']
 
-    train_dataset = datasets.load_dataset("json", data_files=os.path.join(data_source, 'train.json'), split='train')
-    test_dataset = datasets.load_dataset("json", data_files=os.path.join(data_source, 'test.json'), split='train')
+    for data_source in data_sources:
+        train_path = os.path.join(data_source, 'train.json')
+        test_path = os.path.join(data_source, 'test.json')
 
-    train_dataset = train_dataset.map(function=make_map_fn("train", data_source), with_indices=True)
-    test_dataset = test_dataset.map(function=make_map_fn("test", data_source), with_indices=True)
-
-    train_dataset.to_parquet(os.path.join(data_source, 'train.parquet'))
-    test_dataset.to_parquet(os.path.join(data_source, 'test.parquet'))
+        if os.path.exists(train_path):
+            train_dataset = datasets.load_dataset("json", data_files=os.path.join(data_source, 'train.json'), split='train')
+            train_dataset = train_dataset.map(function=make_map_fn("train", data_source), with_indices=True)
+            train_dataset.to_parquet(os.path.join(data_source, 'train.parquet'))
+        
+        if os.path.exists(test_path):
+            test_dataset = datasets.load_dataset("json", data_files=os.path.join(data_source, 'test.json'), split='train')
+            test_dataset = test_dataset.map(function=make_map_fn("test", data_source), with_indices=True)
+            test_dataset.to_parquet(os.path.join(data_source, 'test.parquet'))
+
